+++ src/ballistica/base/support/stdio_console.h
@@ -16,8 +16,9 @@ class StdioConsole {
   auto event_loop() const -> EventLoop* { return event_loop_; }
 
  private:
+  bool use_native_repl{true};
   void StartInMainThread_();
-  void StartNativeREPL_();
+  void StartNativePythonREPL_();
   void PushCommand_(const std::string& command);
   void Clear_();
   EventLoop* event_loop_{};